agrego el seed sólo de primera vez y arreglo permiso de cédulas

+++ database/migrations/2024_10_08_164052_create_usr_roles_table.php
@@ -12,19 +12,22 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('usr_roles', function (Blueprint $table) {
-            $table->id('rol_id');
-            $table->enum('rol_act',['0','1'])->default('1');  ##### Binario de borrado lógico
-            $table->foreignId('rol_usrid')->constrained('users','id'); #### número id del usuario (de table users)
+        // --------------- En producción
+        if(!Schema::hasTable('usr_roles')){
+            Schema::create('usr_roles', function (Blueprint $table) {
+                $table->id('rol_id');
+                $table->enum('rol_act',['0','1'])->default('1');  ##### Binario de borrado lógico
+                $table->foreignId('rol_usrid')->constrained('users','id'); #### número id del usuario (de table users)
 
-            $table->string('rol_crolrol')->default('usr');
-            $table->foreign('rol_crolrol')->references('crol_rol')->on('cat_roles')->onDelete('cascade')->constrained('cat_roles','crol_rol');
+                $table->string('rol_crolrol')->default('usr');
+                $table->foreign('rol_crolrol')->references('crol_rol')->on('cat_roles')->onDelete('cascade')->constrained('cat_roles','crol_rol');
 
-            $table->string('rol_tipo1')->nullable();  ##### En caso de requerirse, descriptor del rol (x ej, con admin-jardin: "JebOax")
-            $table->string('rol_tipo2')->nullable();  ##### En caso de requerirse, descriptor del rol (x ej, con admin-jardin: "animales")
-            $table->string('rol_describe')->nullable()->default('Usuario de sistema');
-            $table->timestamps();
-        });
+                $table->string('rol_tipo1')->nullable();  ##### En caso de requerirse, descriptor del rol (x ej, con admin-jardin: "JebOax")
+                $table->string('rol_tipo2')->nullable();  ##### En caso de requerirse, descriptor del rol (x ej, con admin-jardin: "animales")
+                $table->string('rol_describe')->nullable()->default('Usuario de sistema');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +35,7 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('usr_roles');
+        // --------------- En producción. No borrar porque ya tiene datos de usuarios en producción.
+        // Schema::dropIfExists('usr_roles');
     }
 };
